# feat: enhance cart interface visibility with 40% icon and badge resize