--- a/css/floating-cart.css
+++ b/css/floating-cart.css
@@ -83,16 +83,16 @@
 /* Cart Badge */
 .cart-badge {
   position: absolute;
-  top: -6px;
-  right: -6px;
-  min-width: 20px;
-  height: 20px;
-  padding: 0 5px;
+  top: -8px;
+  right: -8px;
+  min-width: 24px;
+  height: 24px;
+  padding: 0 6px;
   background: var(--cart-primary);
   color: var(--color-white);
   border: 2px solid var(--color-white);
   border-radius: var(--radius-full);
-  font-size: 10px;
+  font-size: 12px;
   font-weight: 700;
   font-family: var(--font-code);
   display: flex;
@@ -492,10 +492,10 @@
   }
 
   .cart-badge {
-    min-width: 18px;
-    height: 18px;
-    padding: 0 4px;
-    font-size: 10px;
+    min-width: 22px;
+    height: 22px;
+    padding: 0 5px;
+    font-size: 11px;
   }
 }
 
@@ -531,10 +531,10 @@
   }
 
   .cart-badge {
-    min-width: 16px;
-    height: 16px;
-    padding: 0 3px;
-    font-size: 9px;
+    min-width: 21px;
+    height: 21px;
+    padding: 0 5px;
+    font-size: 10px;
   }
 }