--- a/css/header-cart.css
+++ b/css/header-cart.css
@@ -33,8 +33,8 @@
 
 /* Cart icon */
 .nav-cart-icon {
-  width: 20px;
-  height: 20px;
+  width: 28px;
+  height: 28px;
   fill: currentColor;
   color: var(--color-text-primary);
   transition: color var(--transition-fast);
@@ -53,15 +53,15 @@
 /* Cart badge in header */
 .nav-cart-badge {
   position: absolute;
-  top: 2px;
-  right: 2px;
-  min-width: 8px;
-  height: 8px;
-  padding: 0 2px;
+  top: 0px;
+  right: 0px;
+  min-width: 11px;
+  height: 11px;
+  padding: 0 3px;
   background: var(--color-red);
   color: var(--color-white);
   border-radius: var(--radius-full);
-  font-size: 9px;
+  font-size: 13px;
   font-weight: 700;
   font-family: var(--font-code);
   display: flex;
@@ -96,17 +96,17 @@
   }
   
   .nav-cart-icon {
-    width: 24px;
-    height: 24px;
+    width: 34px;
+    height: 34px;
   }
-  
+
   .nav-cart-badge {
-    top: -2px;
-    right: -2px;
-    min-width: 10px;
-    height: 10px;
-    font-size: 9px;
-    padding: 0 3px;
+    top: -3px;
+    right: -3px;
+    min-width: 14px;
+    height: 14px;
+    font-size: 13px;
+    padding: 0 4px;
   }
 }