--- a/css/header-cart.css
+++ b/css/header-cart.css
@@ -3,9 +3,10 @@
 /* Cart item in navigation */
 .nav-cart-item {
   position: relative;
-  display: flex;
-  align-items: center;
+  display: inline-flex;
+  align-items: baseline;
   margin-left: 0;
+  vertical-align: baseline;
 }
 
 /* Cart button in header */
@@ -13,12 +14,14 @@
   background: transparent;
   border: none;
   cursor: pointer;
-  padding: var(--space-sm) 0 var(--space-sm) var(--space-sm);
-  display: flex;
+  padding: 0 0 0 var(--space-sm);
+  display: inline-flex;
   align-items: center;
   justify-content: center;
   position: relative;
   transition: transform var(--transition-fast);
+  vertical-align: baseline;
+  line-height: 1;
 }
 
 .nav-cart-button:hover {
@@ -53,15 +56,15 @@
 /* Cart badge in header */
 .nav-cart-badge {
   position: absolute;
-  top: 0px;
-  right: 0px;
+  top: -3px;
+  right: -3px;
   min-width: 11px;
   height: 11px;
   padding: 0 3px;
   background: var(--color-red);
   color: var(--color-white);
   border-radius: var(--radius-full);
-  font-size: 13px;
+  font-size: 10px;
   font-weight: 700;
   font-family: var(--font-code);
   display: flex;
@@ -105,7 +108,7 @@
     right: -3px;
     min-width: 14px;
     height: 14px;
-    font-size: 13px;
+    font-size: 10px;
     padding: 0 4px;
   }
 }
@@ -123,11 +126,11 @@
     margin-left: 0;
   }
 
-  /* Ensure cart icon aligns with nav text */
+  /* Ensure cart icon aligns with nav text baseline */
   .nav-cart-button {
-    height: 100%;
-    display: flex;
-    align-items: center;
+    display: inline-flex;
+    align-items: baseline;
+    vertical-align: baseline;
   }
 }